+++ index.js
@@ -6,8 +6,9 @@ const loginCommon = require('./lib/common/login.js')
 const logoutCommon = require('./lib/common/logout.js')
 const tenant = require('./lib/common/tenant.js')
 const profile = require('./lib/auth0/profile.js')
+const verifyJWT = require('./lib/auth0/verify-jwt.js')
+const auth0ClientFactory = require('./lib/auth0/client-factory.js')
 const getJWT = require('./lib/utils/get-jwt.js')
-const settings = require('./lib/common/settings.js')
 
 const privateVars = new WeakMap()
 
@@ -75,19 +76,12 @@ class BlinkMobileIdentity {
    * Get access token generated after a successful login
    * @returns {String} The access token generated after a successful login.
    */
-  getAccessToken () /* : Promise<string | void> */ {
-    return getJWT()
-  }
-
-  /**
-   * Get settings scoped to a BlinkMobile service.
-   * @param {Object} additionalParameters - Additional parameters to pass to the settings endpoint.
-   * @returns {Object} The settings.
-   */
-  getServiceSettings (
-    additionalParameters /* : Object */
-  ) /* : Promise<Object> */ {
-    return settings((privateVars.get(this) || {}).clientName, additionalParameters)
+  getAccessToken () /* : Promise<string> */ {
+    return Promise.all([
+      getJWT(),
+      auth0ClientFactory.getClientIdByName((privateVars.get(this) || {}).clientName)
+    ])
+      .then(([ jwt, clientId ]) => verifyJWT(jwt, clientId))
   }
 
   /**